+++ lib/src/main/kotlin/com/alibagherifam/kavoshgar/lobby/ServerInformation.kt
@@ -5,7 +5,7 @@ import java.net.InetAddress
 data class ServerInformation(
     val name: String,
     val address: InetAddress,
-    val latency: Int
+    val latency: Long
 ) {
     val addressName: String get() = address.toString().drop(1)
 }
